Splash feature and many things
I find a way to get information more efficient and faster, probably we need to remove mediator because is taking time to load all pokemon in list.

So, first thing is download a light list of pokemon (url, name and id) and save it into database. This process is being done in when splash is displayed.

After that, when user arrives to pokemon list, a shimmer effect will be displayed in each pokemon item when app is getting pokemon details.

- Changed jvm to 18
- Created a splash view model and some classes to fetch initial information (simple pokemon list).
- Added library for splash screen api and coroutines.
- Created splash theme and add some images for launching icon and splash.
- Added extension function test
- Implemented use case pattern